lib: organize methods and complete inline doc
