+++ packages/lib-ts/common/Address.ts
@@ -4,25 +4,36 @@
 // Copyright (c) 2018 Graph Protocol, Inc. and contributors.
 // Modified by Mimic Protocol, 2025.
 
+import { stringToH160 } from '../helpers'
+
 import { ByteArray } from './ByteArray'
 import { Bytes } from './Bytes'
-import { typeConversion } from './conversion'
 
+/**
+ * Represents an Ethereum address, a fixed-length 20-byte value.
+ */
 export class Address extends Bytes {
+  /**
+   * Converts a string representation of an address to an Address instance.
+   */
   static fromString(s: string): Address {
-    return changetype<Address>(typeConversion.stringToH160(s))
+    return changetype<Address>(stringToH160(s))
   }
 
+  /**
+   * Converts a Bytes instance to an Address.
+   * Throws an error if the input is not exactly 20 bytes long.
+   */
   static fromBytes(b: Bytes): Address {
-    if (b.length != 20) {
-      throw new Error(`Bytes of length ${b.length} can not be converted to 20 byte addresses`)
-    }
+    if (b.length != 20) throw new Error(`Bytes of length ${b.length} can not be converted to 20 byte addresses`)
     return changetype<Address>(b)
   }
 
+  /**
+   * Returns a zero address (20 bytes filled with zeroes).
+   */
   static zero(): Address {
     const self = new ByteArray(20)
-
     return changetype<Address>(self)
   }
 }
